[Win] Improve CMake build performance https://bugs.webkit.org/show_bug.cgi?id=155871 <rdar://problem/24747822>
Reviewed by Alex Christensen.

.:

Add a flag to the PROCESS_ALLINONE_FILE macro so that it does not remove
the files contained in the passed all-in-one file, since this breaks
dependency checking and generation of the derived sources from the IDL.
Instead, include the header files in the project so that all files get
generated.

* Source/cmake/WebKitMacros: Updated for 'DerivedSources.cpp' use case.

Source/WebCore:

Treat DerivedSources.cpp as an 'All-in-one' file. Pass a flag to the
PROCESS_ALLINONE_FILE macro so that it does not remove the contents of
the file, since this breaks dependency checking and generation of the
sources from the IDL files. Instead, include the header files in the
project so that all files get generated.

* CMakeLists.txt: Updated for 'DerivedSources.cpp'
* DerivedSources.cpp: Add some generated files that were missing.



Identifier: 173977@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@198669 268f45cc-cd09-0410-ab3c-d52691b4dbfc