Web Automation: add commands to move and resize a browsing context's window https://bugs.webkit.org/show_bug.cgi?id=155349 <rdar://problem/25104911>
Reviewed by Timothy Hatcher.

Parse the new origin or size and request the window to change to the
new frame. This calls through to PageUIClient::setWindowFrame().

* UIProcess/Automation/Automation.json:
Add new enum values for protocol parsing error cases.
Add new commands.

* UIProcess/Automation/WebAutomationSession.cpp:
(WebKit::WebAutomationSession::resizeWindowOfBrowsingContext):
(WebKit::WebAutomationSession::moveWindowOfBrowsingContext):
Added. Parse the incoming payload and bail if nothing would happen
or the values are not usable or out of range. Complain if a change
did not happen when the requested and actual size are different.

* UIProcess/Automation/WebAutomationSession.h:
* UIProcess/WebPageProxy.h: Move setWindowFrame to be public.

Identifier: 174064@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@198760 268f45cc-cd09-0410-ab3c-d52691b4dbfc