CopiedBlock should be 16kB https://bugs.webkit.org/show_bug.cgi?id=156168
Reviewed by Mark Lam.

MarkedBlock is 16kB, and bmalloc's largest fast-path allocation is 16kB,
and the largest page size on Apple devices is 16kB -- so this change
should improve sharing and recycling and keep us on the fast path more.

32kB is also super aggro. At 16kB, we support allocations up to 8kB,
which covers 99.3% of allocations on facebook.com. The 32kB block size
only covered an additional 0.2% of allocations.

* heap/CopiedBlock.h:


Identifier: 174295@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@199016 268f45cc-cd09-0410-ab3c-d52691b4dbfc