REGRESSION(r198955): com.apple.WebKit.Networking.Development crashed in WebKit::NetworkLoad::setPendingDownloadID + 11 https://bugs.webkit.org/show_bug.cgi?id=156177 <rdar://problem/25508037>
Reviewed by Alex Christensen.

We do not currently create a NetworkDataTask when we encounter a BLOB URL when building with
NETWORK_SESSION enabled. This causes us to crash when the download system attempts to work
with a BLOB URL.

* NetworkProcess/NetworkLoad.cpp:
(WebKit::NetworkLoad::setPendingDownloadID): Add null check for m_task.
(WebKit::NetworkLoad::setPendingDownload): Ditto.


Identifier: 174300@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@199021 268f45cc-cd09-0410-ab3c-d52691b4dbfc