Add support for [EnabledAtRuntime] operations on DOMWindow https://bugs.webkit.org/show_bug.cgi?id=156272
Reviewed by Alex Christensen.

Source/JavaScriptCore:

Add identifier for 'fetch' so it can be used from the generated
bindings.

* runtime/CommonIdentifiers.h:

Source/WebCore:

Add support for [EnabledAtRuntime] operations on DOMWindow by omitting
such operations from the static table and add them at run-time in
JSDOMWindow::finishCreation() if the corresponding feature is enabled.

This was needed for window.fetch() for which a hack was temporarily
landed in r199081. This patch drops this hack now that the generated
bindings do the right thing.

* bindings/js/JSDOMGlobalObject.cpp:
(WebCore::JSDOMGlobalObject::scriptExecutionContext):
Drop hack landed in r199081.

* bindings/scripts/CodeGeneratorJS.pm:
(OperationShouldBeOnInstance):
(GeneratePropertiesHashTable):
(GenerateImplementation):
Add support for [EnabledAtRuntime] operations on DOMWindow.


Identifier: 174369@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@199096 268f45cc-cd09-0410-ab3c-d52691b4dbfc