Invalid assertion inside DebuggerScope::getOwnPropertySlot https://bugs.webkit.org/show_bug.cgi?id=156357
Reviewed by Keith Miller.

The Type Profiler might profile JS code that uses DebuggerScope and accesses properties
on it. Therefore, it may have a DebuggerScope object in its log. Objects in the log
are subject to having their getOwnPropertySlot method called. Therefore, the DebuggerScope
might not always be in a valid state when its getOwnPropertySlot method is called.
Therefore, the assertion invalid.

* debugger/DebuggerScope.cpp:
(JSC::DebuggerScope::getOwnPropertySlot):


Identifier: 174451@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@199182 268f45cc-cd09-0410-ab3c-d52691b4dbfc