[css-grid] Remove unnecessary iteration in populateGridPositions loop https://bugs.webkit.org/show_bug.cgi?id=156376
Reviewed by Darin Adler.

The populateGridPositions loop limit was set to 'lastLine'. However, the
the position of last track's start line is updated after the loop, since
it does not follow the same pattern; it does not have a content
distribution offset.

So, since we are essentially overwriting the value stored in the last
iteration, we can just lower the loop limit.

No new tests added, because there is no change in the functionality.

* rendering/RenderGrid.cpp:
(WebCore::RenderGrid::populateGridPositions):


Identifier: 174491@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@199222 268f45cc-cd09-0410-ab3c-d52691b4dbfc